[clang-tidy] Rename AvoidCStyleCastsCheck to AvoidCStyleCastCheck
Make the check name consistent with other modernize checks by using
singular form. Also improve documentation with concrete examples.
