+++ clang-tools-extra/clang-tidy/modernize/AvoidCStyleCastCheck.h
