+++ clang-tools-extra/docs/clang-tidy/checks/modernize/avoid-c-style-cast.rst
@@ -3,12 +3,33 @@
 modernize-avoid-c-style-cast
 ============================
 
-Finds usages of C-style casts.
+This check is similar to ``-Wold-style-cast``, but it suggests automated fixes
+in some cases. The reported locations should not be different from the ones
+generated by ``-Wold-style-cast``.
 
-https://google.github.io/styleguide/cppguide.html#Casting
+C-style casts can perform a variety of different conversions (``const_cast``,
+``static_cast``, ``reinterpret_cast``, or a combination). This makes them
+dangerous as the intent is not clear, and they can silently perform unsafe
+conversions between incompatible types.
+
+
+Examples
+--------
+
+.. code-block:: c++
+
+  class A {
+    public:
+    std::string v;
+  };
+
+  A a;
+  double *num = (double*)(&a);           // Compiles! Hides danger
+  // num = static_cast<double*>(&a);     // Won't compile (good!)
+  num = reinterpret_cast<double*>(&a);   // Compiles, danger is explicit
+
+
+References
+----------
 
 Corresponding cpplint.py check name: `readability/casting`.
-
-This check is similar to ``-Wold-style-cast``, but it suggests automated fixes
-in some cases. The reported locations should not be different from the
-ones generated by ``-Wold-style-cast``.
